lib/aws-batch/requirements.txt
pandas
alpaca-py
boto3
autots